Header Style by cell
Add capability to create custom style for each header, I added variable headerStyles to get the stile on the same way it gets the stiles for the body cells and let us create custom style for each header like on the body I added this variable on the array next to table.styles.headerStyles to be able to get the style when we define the same stile for all the header cells. This is the way I found to make it work since I need to create different halign for each header. 
Please let me know your thoughts,